Split StreamerConnection into twitchEventSubConnection.ts to reduce file complexity
twitchEventSub.ts was still at complexity=52 after the class refactor because
the class body itself counted against the file total. Moving StreamerConnection
(and its helpers isDuplicate, isStale, buildReconnectUrl) into a dedicated file
leaves twitchEventSub.ts with only the thin map-management/public API code.
